move data structures to new folder
