+++ compiler/compiler.c
@@ -3,8 +3,8 @@
 #include "ast.h"
 #include "interpreter/YASL_Object.h"
 #include "middleend.h"
-#include "interpreter/YASL_string.h"
-#include "bytebuffer/bytebuffer.h"
+#include "data-structures/YASL_string.h"
+#include "data-structures/bytebuffer.h"
 #include "parser.h"
 #include "yasl_error.h"
 #include "yasl_include.h"
