+++ data-structures/hash_function.h
@@ -4,7 +4,7 @@
 
 #include "prime/prime.h"
 #include "interpreter/YASL_Object.h"
-#include "interpreter/list.h"
+#include "list.h"
 
 size_t hash_function(const struct YASL_Object s, const size_t a, const size_t m);
 size_t get_hash(const struct YASL_Object s, const size_t num_buckets, const size_t attempt);
