+++ interpreter/VM.c
@@ -6,8 +6,8 @@
 #include <interpreter/YASL_Object.h>
 
 #include "interpreter/builtins.h"
-#include "YASL_string.h"
-#include "hashtable/hashtable.h"
+#include "data-structures/YASL_string.h"
+#include "data-structures/hashtable.h"
 #include "interpreter/refcount.h"
 
 #include "interpreter/table_methods.h"
