+++ interpreter/list_methods.c
@@ -4,7 +4,7 @@
 
 #include "VM.h"
 #include "YASL_Object.h"
-#include "list.h"
+#include "data-structures/list.h"
 #include "yasl_state.h"
 
 int list___get(struct YASL_State *S) {
